+++ lib/app/components/show_vehicle_component.dart
@@ -1,4 +1,5 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:motapp/app/model/vehicle_model.dart';
 import 'package:motapp/app/pages/vehicles/register_vehicle_page.dart';
@@ -81,21 +82,27 @@ class _ShowVehicleComponentState extends State<ShowVehicleComponent> {
                               ),
                               TextButton(
                                 onPressed: () {
-                                  ScaffoldMessenger.of(context).showSnackBar(
-                                    SnackBar(
-                                      content: Text(
-                                        'Veículo ${vehicles.placa} excluído com sucesso!',
-                                        style: TextStyle(
-                                          fontSize: 18,
-                                          fontWeight: FontWeight.bold,
+                                  final userId =
+                                      FirebaseAuth.instance.currentUser?.uid;
+                                  if (userId != null) {
+                                    ScaffoldMessenger.of(context).showSnackBar(
+                                      SnackBar(
+                                        content: Text(
+                                          'Veículo ${vehicles.placa} excluído com sucesso!',
+                                          style: TextStyle(
+                                            fontSize: 18,
+                                            fontWeight: FontWeight.bold,
+                                          ),
                                         ),
                                       ),
-                                    ),
-                                  );
-                                  FirebaseFirestore.instance
-                                      .collection('veiculos')
-                                      .doc(vehicles.id)
-                                      .delete();
+                                    );
+                                    FirebaseFirestore.instance
+                                        .collection('usuarios')
+                                        .doc(userId)
+                                        .collection('veiculos')
+                                        .doc(vehicles.id)
+                                        .delete();
+                                  }
                                   Navigator.of(context).pop();
                                 },
                                 child: const Text(
